feat: ajout tracking plausible sur page de stats
