+++ ui/pages/stats/index.jsx
@@ -1,0 +1,3 @@
+import StatistiquesPage from "./[tab]";
+
+export default StatistiquesPage;
